+++ src/main/java/online/kingdomkeys/kingdomkeys/lib/Strings.java
@@ -74,8 +74,6 @@ public class Strings {
 
 		Magic_DarkFiraga = KingdomKeys.MODID + ":" + Mag_Prefix + "darkfiraga",
 		Magic_TripleFiraga = KingdomKeys.MODID + ":" + Mag_Prefix + "triplefiraga",
-		Magic_Blackout = KingdomKeys.MODID + ":" + Mag_Prefix + "blackout",
-		Magic_Poison = KingdomKeys.MODID + ":" + Mag_Prefix + "posion",
 		Magic_CrawlingFiraga = KingdomKeys.MODID + ":" + Mag_Prefix + "crawlingfiraga",
 		Magic_FissionFiraga = KingdomKeys.MODID + ":" + Mag_Prefix + "fissionfiraga",
 		Magic_FiragaBurst = KingdomKeys.MODID + ":" + Mag_Prefix + "firagaburst",
@@ -83,10 +81,14 @@ public class Strings {
 		Magic_TripleBlizzard = KingdomKeys.MODID + ":" + Mag_Prefix + "tripleblizzaga",
 		Magic_DeepFreeze = KingdomKeys.MODID + ":" + Mag_Prefix + "deepfreeze",
 		Magic_Glacier = KingdomKeys.MODID + ":" + Mag_Prefix + "glacier",
-		Magic_IceBarrage = KingdomKeys.MODID + ":" + Mag_Prefix + "icebarrage";
-																
-    
-    // Drive Forms
+		Magic_IceBarrage = KingdomKeys.MODID + ":" + Mag_Prefix + "icebarrage",
+
+		Magic_ThundagaShot = KingdomKeys.MODID + ":" + Mag_Prefix + "thundagashot",
+
+		Magic_Blackout = KingdomKeys.MODID + ":" + Mag_Prefix + "blackout",
+		Magic_Poison = KingdomKeys.MODID + ":" + Mag_Prefix + "posion";
+
+	// Drive Forms
     public static final String
     		DF_Prefix = "form_",
             Form_Valor = KingdomKeys.MODID+":"+DF_Prefix+"valor",
@@ -691,6 +693,8 @@ public class Strings {
 		SpellDeepFreeze = "spelldeepfreeze",
 		SpellGlacier = "spellglacier",
 		SpellIceBarrage = "spellicebarrage",
+
+		SpellThundagaShot = "spellthundagashot",
 
 		StruggleSword = "struggle_sword",
 		StruggleWand = "struggle_wand",